refactor App files
